Moved insert metrics to auth.js
